requirements.txt
click==8.1.7
python-dotenv==1.0.1
colorama==0.4.6
